Move implementation of dependency interface to primary providers
