+++ dependency/main.go
@@ -5,7 +5,8 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/konveyor/analyzer-lsp/dependency/provider/java"
+	"github.com/konveyor/analyzer-lsp/provider/java"
+	"github.com/konveyor/analyzer-lsp/provider/lib"
 	"gopkg.in/yaml.v2"
 )
 
@@ -16,7 +17,7 @@ var (
 func main() {
 	flag.Parse()
 
-	p := java.GetDepProvider()
+	p := java.NewJavaProvider(lib.Config{Location: "../examples/java"})
 	deps, err := p.GetDependencies("../examples/java/pom.xml")
 
 	if err != nil {
