+++ provider/java/dependency_test.go
